The project's objectives were to automate the bank's intern management system and enhance its efficacy.
The group thoroughly examined the existing intern management system and made suggestions for its automation. The adoption of an MS SQL database system, the development of a centralized intern monitoring unit, and the creation of open lines of communication between the various departments involved in the process were all praised.